app/manager/requirements.txt
requests
Flask
fpdf2>=2.7